// src/api/subscription/api.ts
export const SubscriptionHistoryResponseStatusEnum = {
  Active: 'ACTIVE',
  Canceled: 'CANCELED',
  Incomplete: 'INCOMPLETE',
  IncompleteExpired: 'INCOMPLETE_EXPIRED',
  PastDue: 'PAST_DUE',
  Trialing: 'TRIALING',
  Unpaid: 'UNPAID',
} as const;

export type SubscriptionHistoryResponseStatusEnum =
  (typeof SubscriptionHistoryResponseStatusEnum)[keyof typeof SubscriptionHistoryResponseStatusEnum];